pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spotify-lifecycle"
version = "0.1.0"
description = "Serverless, event-driven Spotify Lifecycle Manager"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "spotipy>=2.22.1",
    "boto3>=1.28.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.0.292",
    "black>=23.7.0",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I001"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["spotify_lifecycle"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v"

[dependency-groups]
dev = [
    "black>=25.12.0",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]
